+++ Assets/LDtkUnity/Editor/CustomEditor/SectionElementDrawers/LDtkDrawerEntity.cs
@@ -1,4 +1,5 @@
-﻿using UnityEditor;
+﻿using System.Linq;
+using UnityEditor;
 using UnityEngine;
 
 namespace LDtkUnity.Editor
@@ -7,6 +8,30 @@ namespace LDtkUnity.Editor
     {
         public LDtkDrawerEntity(EntityDefinition def, SerializedProperty obj, string key) : base(def, obj, key)
         {
+        }
+
+        public override void Draw()
+        {
+            Rect controlRect = EditorGUILayout.GetControlRect();
+
+            GUIContent objectContent = new GUIContent()
+            {
+                text = _data.Identifier,
+            };
+
+            if (!_data.FieldDefs.IsNullOrEmpty())
+            {
+                objectContent.tooltip = string.Join("\n", _data.FieldDefs.Select(p => p.Identifier));
+            }
+            
+            EditorGUI.PropertyField(controlRect, Value, objectContent);
+            
+            
+            if (HasProblem())
+            {
+                DrawCachedProblem(controlRect);
+            }
+
         }
 
         protected override string AssetUnassignedText => "No prefab assigned; Entity instance won't show up in the import result";